+++ src/ui/mod.rs
@@ -3,6 +3,7 @@
 pub(crate) mod channel_sidebar;
 pub(crate) mod device_sidebar;
 pub(crate) mod exact_discovery_dialog;
+pub(crate) mod icons;
 pub(crate) mod idle_inhibitor;
 pub(crate) mod objects;
 pub(crate) mod player_view;